requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.14.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2